Add support for uploading HLS segments and playlists (#41)
This adds support for uploading segments and playlists when generating segmented files.

A segment and the corresponding playlist update are uploaded as soon as they are ready. Uploads are serialized on a single queue to make sure they occur in order, to prevent overwriting a playlist with an older one. This refactors some of pipeline.go and upload.go to be able to reuse the existing upload code.