Push handler metrics to main service when exiting (#565)
This PR prevents the last prometheus metrics to be lost at the end of an egress by pushing them in a new ipc message to the main service process. This is particularly important since the last set of metrics contains file and manifest upload information.

The metrics pushed by the handler on exit are stored in an array in the Service object. They are appended to the other metrics in the prometheus gatherer.