Use specific template version with egress versions (#1000)
This changes the egress build system to refer to a specific egress-template docker image tag instead of "latest". This allows pinning an egress build to a specific template version instead of the current scheme where the latest is always used.

In order to do this:
- the publish-template action injects the image tag into version.go and commits the result to the topic branch
- this image version is read by the magefile and the other GitHub Actions and passed into the docker build calls.

This means that version.go is currently partly manually maintained and partly autogenerated. There is also a new utility that just prints the template version, for use by GitHub Actions (including eventually by cloud-egress builds).