Fix for racy addiion of app sources (#1073)
After BuildPipeline has called p.Link() and returned, but before p.Run() moves the state out of StateBuilding - if we get track added and start adding source in between - there will be nothing to link it afterwards.
As a consequence all tracks unlucky enough to land during that gap would be lost, source wouldn't be linked but since added to the pipeline it's state will regularly change state to Playing. All packets appwriters puhes will be discarded as the branch isn't linked.
The fix is to upgrade state to started just after linking static parts of the pipeline but before closing build ready fuse.