Ensure pipeline is at least in PAUSED state before pushing data from app sources
In very rare occasions it could happen that app sources start pushing packets before all pads are activated which causes the appsrc pad task to switch to paused state - preventing any future flow, effectively dropping all packets for the affected track.
The issue might be due to pipeline not reaching at least PAUSED state before sources are in PLAYING state.
The main goal of this change is to ensure app writers are waiting until the gst pipeline is in PAUSED state.