Drain appwriter (#1090)
* Drain appwriter

When Drain() was called on an AppWriter, not all samples buffered in the jitter buffer and internal channel were being pushed to the pipeline before the stream ended. This resulted possibly lost audio/video samples at the end of tracks.

Replaced the single endStream fuse with three separate fuses to track different stages of the drain process:
  - endStreamSignaled: Drain/end has been requested
  - endStreamSourceProcessed: Jitter buffer has been flushed and all packets extracted
  - endStreamProcessed: All samples have been pushed to the pipeline

Added new Flush API on jitter buffer to drain the remaining samples as well at the end.

* timeout on wating for eos fuse to be broken